pipeline/requirements.txt
whisperx
pyannote.audio
torch
torchaudio
numpy
scipy
